Move CachedDownloadableItem to BlobbySerializer
Moves `CachedDownloadableItem` to the `BlobbySerializer` file, since
that's the only place it's used.
