+++ extensions/data-transfer/portability-data-transfer-generic/src/main/java/org/datatransferproject/datatransfer/generic/BlobbySerializer.java
@@ -10,25 +10,45 @@ import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Queue;
+import org.datatransferproject.types.common.DownloadableItem;
 import org.datatransferproject.types.common.models.blob.BlobbyStorageContainerResource;
 import org.datatransferproject.types.common.models.blob.DigitalDocumentWrapper;
 import org.datatransferproject.types.common.models.blob.DtpDigitalDocument;
 
-class BlobbyContainerPath {
-  private BlobbyStorageContainerResource container;
-  private String path;
+/**
+ * Wrapper to adapt items known to be in temp storage (e.g. BLOB data) into {@link DownloadableItem}
+ *
+ * <p>It's useful to wrap such items so upstream code can consume either known temp store'd items or
+ * items the Importer has to download itself (some MEDIA items) from the same interface.
+ */
+class CachedDownloadableItem implements DownloadableItem {
+  private String cachedId;
+  private String name;
 
-  public BlobbyContainerPath(BlobbyStorageContainerResource container, String path) {
-    this.container = container;
-    this.path = path;
+  public CachedDownloadableItem(String cachedId, String name) {
+    this.cachedId = cachedId;
+    this.name = name;
   }
 
-  public BlobbyStorageContainerResource getContainer() {
-    return container;
+  @Override
+  public String getIdempotentId() {
+    return cachedId;
   }
 
-  public String getPath() {
-    return path;
+  @Override
+  public String getFetchableUrl() {
+    // 'url' is ID when cached
+    return cachedId;
+  }
+
+  @Override
+  public boolean isInTempStore() {
+    return true;
+  }
+
+  @Override
+  public String getName() {
+    return name;
   }
 }
 
@@ -67,6 +87,24 @@ class BlobbyFolder {
 }
 
 public class BlobbySerializer {
+  static class BlobbyContainerPath {
+    private BlobbyStorageContainerResource container;
+    private String path;
+
+    public BlobbyContainerPath(BlobbyStorageContainerResource container, String path) {
+      this.container = container;
+      this.path = path;
+    }
+
+    public BlobbyStorageContainerResource getContainer() {
+      return container;
+    }
+
+    public String getPath() {
+      return path;
+    }
+  }
+
   static final String SCHEMA_SOURCE =
       GenericTransferConstants.SCHEMA_SOURCE_BASE
           + "/extensions/data-transfer/portability-data-transfer-generic/src/main/java/org/datatransferproject/datatransfer/generic/BlobbySerializer.java";
